Make all sphinx extensions except bokeh-plot be internal
