+++ docs/bokeh/source/docs/reference/sphinxext.rst
@@ -3,113 +3,9 @@
 bokeh.sphinxext
 ===============
 
-In order to help automate and simplify the generation of Bokeh
-documentation, several Sphinx extensions have been created. Most
-of these will not be of general interest, except for `bokeh.sphinxext.bokeh_plot`_,
-which allows anyone to include and embed bokeh plots directly
-in their own Sphinx docs.
-
-.. _bokeh.sphinxext.bokeh_autodoc:
-
-bokeh_autodoc
--------------
-
-.. automodule:: bokeh.sphinxext.bokeh_autodoc
-
-.. _bokeh.sphinxext.bokeh_color:
-
-bokeh_color
------------
-
-.. automodule:: bokeh.sphinxext.bokeh_color
-
-.. _bokeh.sphinxext.bokeh_dataframe:
-
-bokeh_dataframe
----------------
-
-.. automodule:: bokeh.sphinxext.bokeh_dataframe
-
-.. _bokeh.sphinxext.bokeh_enum:
-
-bokeh_enum
-----------
-
-.. automodule:: bokeh.sphinxext.bokeh_enum
-
-.. _bokeh.sphinxext.bokeh_gallery:
-
-bokeh_gallery
--------------
-
-.. automodule:: bokeh.sphinxext.bokeh_gallery
-
-.. _bokeh.sphinxext.bokeh_jinja:
-
-bokeh_jinja
------------
-
-.. automodule:: bokeh.sphinxext.bokeh_jinja
-
-.. _bokeh.sphinxext.bokeh_model:
-
-bokeh_model
------------
-
-.. automodule:: bokeh.sphinxext.bokeh_model
-
-.. _bokeh.sphinxext.bokeh_options:
-
-bokeh_options
--------------
-
-.. automodule:: bokeh.sphinxext.bokeh_options
-
-.. _bokeh.sphinxext.bokeh_palette:
-
-bokeh_palette
--------------
-
-.. automodule:: bokeh.sphinxext.bokeh_palette
-
-.. _bokeh.sphinxext.bokeh_palette_group:
-
-bokeh_palette_group
--------------------
-
-.. automodule:: bokeh.sphinxext.bokeh_palette_group
-
-.. _bokeh.sphinxext.bokeh_plot:
+Sphinx extensions for including Bokeh content in Sphinx documentation.
 
 bokeh_plot
 ----------
 
 .. automodule:: bokeh.sphinxext.bokeh_plot
-
-.. _bokeh.sphinxext.bokeh_prop:
-
-bokeh_prop
-----------
-
-.. automodule:: bokeh.sphinxext.bokeh_prop
-
-.. _bokeh.sphinxext.bokeh_releases:
-
-bokeh_releases
---------------
-
-.. automodule:: bokeh.sphinxext.bokeh_releases
-
-.. _bokeh.sphinxext.bokeh_roles:
-
-bokeh_roles
------------
-
-.. automodule:: bokeh.sphinxext.bokeh_roles
-
-.. _bokeh.sphinxext.bokeh_sitemap:
-
-bokeh_sitemap
--------------
-
-.. automodule:: bokeh.sphinxext.bokeh_sitemap
