+++ src/bokeh/core/property/bases.py
@@ -88,9 +88,7 @@ class Property(PropertyDescriptorFactory[T]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     """
