+++ src/bokeh/core/property/json.py
@@ -51,9 +51,7 @@ class JSON(String):
             have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     """
