+++ src/bokeh/core/property/numeric.py
@@ -166,9 +166,7 @@ class Size(Float):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
 
@@ -207,9 +205,7 @@ class Percent(Float):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
 
@@ -253,9 +249,7 @@ class Angle(Float):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     """
