+++ src/bokeh/core/property/primitive.py
@@ -70,9 +70,7 @@ class Bool(PrimitiveProperty[bool]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
 
@@ -105,9 +103,7 @@ class Complex(PrimitiveProperty[complex]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     """
 
@@ -124,9 +120,7 @@ class Int(PrimitiveProperty[int]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
 
@@ -159,9 +153,7 @@ class Float(PrimitiveProperty[float]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
 
@@ -205,9 +197,7 @@ class String(PrimitiveProperty[str]):
             A default value for attributes created from this property to have.
 
         help (str or None, optional) :
-            A documentation string for this property. It will be automatically
-            used by the :ref:`bokeh.sphinxext.bokeh_prop` extension when
-            generating Spinx documentation. (default: None)
+            A documentation string for this property. (default: None)
 
     Example:
