+++ src/bokeh/sphinxext/_internal/_templates/bokehjs_content_prologue.html
