+++ src/bokeh/sphinxext/_internal/_templates/color_detail.html
