+++ src/bokeh/sphinxext/_internal/_templates/enum_detail.rst
