+++ src/bokeh/sphinxext/_internal/_templates/example_metadata.rst
