+++ src/bokeh/sphinxext/_internal/_templates/gallery_detail.rst
