+++ src/bokeh/sphinxext/_internal/_templates/gallery_page.rst
