+++ src/bokeh/sphinxext/_internal/_templates/jinja_detail.rst
