+++ src/bokeh/sphinxext/_internal/_templates/model_detail.rst
