+++ src/bokeh/sphinxext/_internal/_templates/options_detail.rst
