+++ src/bokeh/sphinxext/_internal/_templates/palette_detail.html
