+++ src/bokeh/sphinxext/_internal/_templates/palette_group_detail.html
