+++ src/bokeh/sphinxext/_internal/_templates/prop_detail.rst
