+++ src/bokeh/sphinxext/_internal/_templates/release_detail.rst
