+++ src/bokeh/sphinxext/_internal/_templates/settings_detail.rst
