+++ src/bokeh/sphinxext/_internal/bokeh_directive.py
@@ -4,8 +4,7 @@
 #
 # The full license is in the file LICENSE.txt, distributed with this software.
 # -----------------------------------------------------------------------------
-""" Install some functions for the bokeh theme to make use of.
-
+""" Provide a base class and useful functions for Bokeh Sphinx directives.
 
 """
 
@@ -13,10 +12,6 @@
 # Boilerplate
 # -----------------------------------------------------------------------------
 from __future__ import annotations
-
-# Standard library imports
-import os
-from pathlib import PurePath
 
 import logging  # isort:skip
 
@@ -26,15 +21,33 @@ log = logging.getLogger(__name__)
 # Imports
 # -----------------------------------------------------------------------------
 
-# Bokeh imports
-from bokeh.resources import Resources
-from bokeh.settings import settings
+# Standard library imports
+import re
+
+# External imports
+from docutils import nodes
+from docutils.statemachine import ViewList
+from sphinx.util.docutils import SphinxDirective
+from sphinx.util.nodes import nested_parse_with_titles
 
 # -----------------------------------------------------------------------------
 # Globals and constants
 # -----------------------------------------------------------------------------
 
-__all__ = ("get_sphinx_resources",)
+# taken from Sphinx autodoc
+py_sig_re = re.compile(
+    r"""^ ([\w.]*\.)?            # class name(s)
+          (\w+)  \s*             # thing name
+          (?: \((.*)\)           # optional: arguments
+           (?:\s* -> \s* (.*))?  # return annotation
+          )? $                   # and nothing more
+          """,
+    re.VERBOSE,
+)
+
+__all__ = (
+    "BokehDirective",
+)
 
 # -----------------------------------------------------------------------------
 # General API
@@ -44,32 +57,18 @@ __all__ = ("get_sphinx_resources",)
 # Dev API
 # -----------------------------------------------------------------------------
 
-# TODO (bv) this needs to be a configuration
-_REPO_TOP = PurePath(os.getcwd()).parents[1]
 
-def get_sphinx_resources(include_bokehjs_api=False):
-    docs_cdn = settings.docs_cdn()
+class BokehDirective(SphinxDirective):
 
-    # if BOKEH_DOCS_CDN is unset just use default CDN resources
-    if docs_cdn is None:
-        resources = Resources(mode="cdn")
-    else:
-        # "BOKEH_DOCS_CDN=local" is used for building and displaying the docs locally
-        if docs_cdn == "local":
-            resources = Resources(mode="server", root_url="/en/latest/")
+    def parse(self, rst_text, annotation):
+        result = ViewList()
+        for line in rst_text.split("\n"):
+            result.append(line, annotation)
+        node = nodes.paragraph()
+        node.document = self.state.document
+        nested_parse_with_titles(self.state, result, node)
+        return node.children
 
-        # "BOKEH_DOCS_CDN=test:newthing" is used for building and deploying test docs to
-        # a one-off location "en/newthing" on the docs site
-        elif docs_cdn.startswith("test:"):
-            version = docs_cdn.split(":")[1]
-            resources = Resources(mode="server", root_url=f"/en/{version}/")
-
-        # Otherwise assume it is a dev/rc/full release version and use CDN for it
-        else:
-            resources = Resources(mode="cdn", version=docs_cdn)
-    if include_bokehjs_api:
-        resources.components.append("bokeh-api")
-    return resources
 
 # -----------------------------------------------------------------------------
 # Private API
