+++ src/bokeh/sphinxext/_internal/bokeh_enum.py
@@ -15,36 +15,9 @@ module name as an option. An optional description may be added as content:
 
 .. code-block:: rest
 
-    .. bokeh-enum:: baz
-        :module: bokeh.sphinxext.sample
+    .. bokeh-enum:: some_enum
 
-        Specify a baz style
-
-Examples
---------
-
-The directive above will generate the following output:
-
-=====
-
-.. bokeh-enum:: baz
-    :module: bokeh.sphinxext.sample
-
-    Specify a baz style
-
-=====
-
-Although ``bokeh-enum`` may be used explicitly, it is more often convenient in
-conjunction with the :ref:`bokeh.sphinxext.bokeh_autodoc` extension. Together,
-the same output above will be generated directly from the following code:
-
-.. code-block:: python
-
-    #: Specify a baz style
-    baz = enumeration("a", "b", "c")
-
-To enable this extension, add `"bokeh.sphinxext.bokeh_enum"` to the extensions
-list in your Sphinx configuration module.
+        Specify a some_enum value
 
 """
