+++ src/bokeh/sphinxext/_internal/bokeh_palette.py
@@ -64,11 +64,6 @@ Will generate the output:
 
 :bokeh-palette:`viridis(256)`
 
------
-
-To enable this extension, add `"bokeh.sphinxext.bokeh_palette"` to the
-extensions list in your Sphinx configuration module.
-
 """
 
 # -----------------------------------------------------------------------------
