+++ src/bokeh/sphinxext/_internal/bokeh_roles.py
@@ -38,9 +38,6 @@ The repo history shows that :bokeh-commit:`bf19bcb` was made in
 in :bokeh-pull:`1698`,which closed :bokeh-issue:`1694`. This included
 updating all of the files in the :bokeh-tree:`examples` subdirectory.
 
-To enable this extension, add `"bokeh.sphinxext.bokeh_roles"` to the
-extensions list in your Sphinx configuration module.
-
 """
 
 # -----------------------------------------------------------------------------
@@ -65,7 +62,7 @@ from docutils import nodes, utils
 from docutils.parsers.rst.roles import set_classes
 
 # Bokeh imports
-from . import PARALLEL_SAFE
+from . import PARALLEL_SAFE, REPO_TOP
 
 # -----------------------------------------------------------------------------
 # Globals and constants
@@ -134,7 +131,6 @@ def bokeh_minpy(name, rawtext, text, lineno, inliner, options=None, content=None
     empty.
 
     """
-    from ._internal import REPO_TOP
     pyproject = toml.load(join(REPO_TOP, "pyproject.toml"))
     node = nodes.Text(pyproject["project"]["requires-python"].lstrip(">="))
     return [node], []
@@ -169,7 +165,6 @@ def bokeh_requires(name, rawtext, text, lineno, inliner, options=None, content=N
     empty.
 
     """
-    from ._internal import REPO_TOP
     pyproject = toml.load(join(REPO_TOP, "pyproject.toml"))
     node = nodes.bullet_list()
     for dep in pyproject["project"]["dependencies"]:
