+++ src/bokeh/sphinxext/_internal/bokeh_sampledata_xref.py
@@ -13,9 +13,6 @@ The ``bokeh-sampledata-xref`` directive can be used by supplying:
 This can be used to add links to all existing standalone examples in the
 documentation.
 
-To enable this extension, add `"bokeh.sphinxext.bokeh_sampledata_xref"` to the
-extensions list in your Sphinx configuration module.
-
 """
 
 #-----------------------------------------------------------------------------
