+++ src/bokeh/sphinxext/_internal/example_handler.py
@@ -4,9 +4,6 @@
 #
 # The full license is in the file LICENSE.txt, distributed with this software.
 # -----------------------------------------------------------------------------
-"""
-
-"""
 
 # -----------------------------------------------------------------------------
 # Boilerplate
